add LottoFactory
