+++ src/main/java/lotto/Lotto.java
@@ -6,10 +6,10 @@ import lotto.exception.IllegalLottoNumberFormatException;
 import java.util.*;
 
 public class Lotto {
-    private static final int LOTTO_SIZE = 6;
+    protected static final int LOTTO_SIZE = 6;
     private List<LottoNumber> lottoTicket;
 
-    private Lotto(List<LottoNumber> lottoTicket) {
+    protected Lotto(List<LottoNumber> lottoTicket) {
         if (lottoTicket == null || lottoTicket.isEmpty()) {
             throw new IllegalArgumentException();
         }
@@ -19,47 +19,6 @@ public class Lotto {
         }
 
         this.lottoTicket = lottoTicket;
-    }
-
-    public static Lotto makeAutoLottoNumber() {
-        Set<LottoNumber> lottoNumber = new HashSet<>();
-        while (lottoNumber.size() < LOTTO_SIZE) {
-            lottoNumber.add(LottoNumber.getLottoNumber(randomNumber()));
-        }
-        return new Lotto(setToList(lottoNumber));
-    }
-
-    private static Integer randomNumber() {
-        return (int) (Math.random() * 45) + 1;
-    }
-
-    public static List<Lotto> makeManualLottoNumbers(List<String> inputLottoNumbers) throws IllegalArgumentException {
-        List<Lotto> manualLottoNumbers = new ArrayList<>();
-        for(String inputLottoNumber : inputLottoNumbers) {
-            manualLottoNumbers.add(makeLottoNumber(stringToList(inputLottoNumber)));
-        }
-        return manualLottoNumbers;
-    }
-
-    public static Lotto makeManualLottoNumber(String inputLottoNumber) throws IllegalArgumentException {
-        return makeLottoNumber(stringToList(inputLottoNumber));
-    }
-
-    public static Lotto makeLottoNumber(List<String> manualLottoNumber) throws IllegalArgumentException {
-        Set<LottoNumber> lottoNumber = new HashSet<>();
-        for (String number : manualLottoNumber) {
-            lottoNumber.add(LottoNumber.getLottoNumber(number));
-        }
-        return new Lotto(setToList(lottoNumber));
-    }
-
-    public static List<LottoNumber> setToList(Set<LottoNumber> lottoNumber) {
-        return sort(new ArrayList<>(lottoNumber));
-    }
-
-    public static List<LottoNumber> sort(List<LottoNumber> lottoNumber) {
-        Collections.sort(lottoNumber);
-        return lottoNumber;
     }
 
     public int checkDuplicateBonusNumber(int bonus) {
@@ -87,21 +46,6 @@ public class Lotto {
 
     public String toString() {
         return lottoTicket.toString();
-    }
-
-    private static List<String> stringToList(String lottoNumber) {
-        return split(lottoNumber);
-    }
-
-    private static List<String> split(String winnerNumber) {
-        return Arrays.asList(trim(winnerNumber.split(",")));
-    }
-
-    private static String[] trim(String[] winnerNumbers) {
-        for (String winnerNumber : winnerNumbers) {
-            winnerNumber.trim();
-        }
-        return winnerNumbers;
     }
 
 }
